Implemented BarChart to visually display current weight vs target goal weight.
Added functionality to update the bar chart when weight or goal changes.

Got a text view to display the remaining weight to reach the users goal.

Updated bar chart colors to differentiate between weight (green) and goal (pink).

Integrated weight data from ViewModel to update chart and show user progress.

Tested to see if the bar chart changes to correct amount when updating the weight and goal and it does.

Changed the menu color to the app background for a nicer user experience.